Move NPCs dialogs to their own files
