+++ apps/freeablo/fagui/npcs/pepindialog.h
@@ -1,0 +1,18 @@
+#pragma once
+
+#include "../characterdialoguepopup.h"
+#include "../../faworld/actor.h"
+
+namespace FAGui
+{
+    class PepinDialog : public CharacterDialoguePopup
+    {
+    public:
+        PepinDialog(GuiManager& guiManager, FAWorld::Actor* actor);
+
+    protected:
+        virtual DialogData getDialogData() override;
+
+        FAWorld::Actor* mActor = nullptr;
+    };
+}
